+++ MatomoTracker.php
@@ -28,6 +28,7 @@ class MatomoTracker
      * MatomoTracker::$URL = 'http://yourwebsite.org/matomo/';
      *
      * @var string
+     * @deprecated
      */
     static public $URL = '';
 
@@ -202,6 +203,8 @@ class MatomoTracker
 
     private $requestMethod = null;
 
+    private string $apiUrl = '';
+
     /**
      * Builds a MatomoTracker object, used to track visits, pages and Goal conversions
      * for a specific website, by using the Matomo Tracking API.
@@ -227,6 +230,7 @@ class MatomoTracker
         );
         if (!empty($apiUrl)) {
             self::$URL = $apiUrl;
+            $this->apiUrl = $apiUrl;
         }
 
         $this->setNewVisitorId();
@@ -240,6 +244,7 @@ class MatomoTracker
     public function setApiUrl(string $url): void
     {
         self::$URL = $url;
+        $this->apiUrl = $url;
     }
 
     /**
@@ -2152,23 +2157,26 @@ didn't change any existing VisitorId value */
 
     /**
      * Returns the base URL for the Matomo server.
+     *
+     * @throws Exception
      */
     protected function getBaseUrl(): string
     {
-        if (empty(self::$URL)) {
+        if ($this->apiUrl === '') {
             throw new Exception(
                 'You must first set the Matomo Tracker URL by calling
                  MatomoTracker::$URL = \'http://your-website.org/matomo/\';'
             );
         }
-        if (strpos(self::$URL, '/matomo.php') === false
-            && strpos(self::$URL, '/proxy-matomo.php') === false
+
+        if (str_contains($this->apiUrl, '/matomo.php') === false
+            && str_contains($this->apiUrl, '/proxy-matomo.php') === false
         ) {
-            self::$URL = rtrim(self::$URL, '/');
-            self::$URL .= '/matomo.php';
+            $this->apiUrl = rtrim($this->apiUrl, '/');
+            $this->apiUrl .= '/matomo.php';
         }
 
-        return self::$URL;
+        return $this->apiUrl;
     }
 
     /**